rename tomography_utils.py to cdi_utils.py
